+++ src/cqportal/ContainerFrame.py
@@ -1,0 +1,12 @@
+import cadquery as cq
+from . import Frame
+
+class ContainerFrame(Frame):
+    def __init__(self):
+        super().__init__()
+        
+    def make(self, parent=None):
+        super().make(parent)
+            
+    def build(self):
+        return super().build()